Fixing launching
