+++ src/MainFrame.java
@@ -125,11 +125,14 @@ public class MainFrame extends JFrame {
         }
     }
 
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(new Runnable() {
-            public void run() {
-                new MainFrame();
-            }
-        });
+    public static class Main {
+        public static void main(String[] args) {
+            SwingUtilities.invokeLater(new Runnable() {
+                public void run() {
+                    new MainFrame();
+                }
+            });
+        }
+
     }
 }
